api/requirements.txt
fastapi
uvicorn
pyttsx3
ollama
numpy
python-multipart
requests
websocket-client


